Clean up route comparison
