+++ internal/k8s/reconciler/listener_test.go
@@ -9,13 +9,11 @@ import (
 	"github.com/stretchr/testify/require"
 	k8s "k8s.io/api/core/v1"
 	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	gw "sigs.k8s.io/gateway-api/apis/v1alpha2"
 
 	"github.com/hashicorp/consul-api-gateway/internal/core"
 	"github.com/hashicorp/consul-api-gateway/internal/k8s/gatewayclient/mocks"
 	"github.com/hashicorp/consul-api-gateway/internal/store"
-	storeMocks "github.com/hashicorp/consul-api-gateway/internal/store/mocks"
 	"github.com/hashicorp/go-hclog"
 )
 
@@ -352,19 +350,6 @@ func TestListenerConfig(t *testing.T) {
 	}).Config())
 }
 
-func TestRouteAddedCallbacks(t *testing.T) {
-	t.Parallel()
-
-	listener := NewK8sListener(&K8sGateway{Gateway: &gw.Gateway{}}, gw.Listener{}, K8sListenerConfig{
-		Logger: hclog.NewNullLogger(),
-	})
-	require.Equal(t, int32(0), listener.ListenerState.RouteCount)
-	listener.OnRouteAdded(nil)
-	require.Equal(t, int32(1), listener.ListenerState.RouteCount)
-	listener.OnRouteRemoved("")
-	require.Equal(t, int32(0), listener.ListenerState.RouteCount)
-}
-
 func TestListenerStatusConditions(t *testing.T) {
 	t.Parallel()
 
@@ -372,338 +357,4 @@ func TestListenerStatusConditions(t *testing.T) {
 		Logger: hclog.NewNullLogger(),
 	})
 	require.Len(t, listener.Status().Conditions, 4)
-}
-
-func TestListenerCanBind(t *testing.T) {
-	t.Parallel()
-
-	// alternative type
-	listener := NewK8sListener(&K8sGateway{Gateway: &gw.Gateway{}}, gw.Listener{}, K8sListenerConfig{
-		Logger: hclog.NewNullLogger(),
-	})
-	canBind := listener.Bind(context.Background(), storeMocks.NewMockRoute(nil))
-	require.False(t, canBind)
-
-	factory := NewFactory(FactoryConfig{
-		Logger: hclog.NewNullLogger(),
-	})
-
-	// no match
-	listener = NewK8sListener(&K8sGateway{Gateway: &gw.Gateway{}}, gw.Listener{}, K8sListenerConfig{
-		Logger: hclog.NewNullLogger(),
-	})
-	canBind = listener.Bind(context.Background(), factory.NewRoute(&gw.HTTPRoute{}))
-	require.False(t, canBind)
-
-	// match
-	listener = NewK8sListener(&K8sGateway{
-		Gateway: &gw.Gateway{
-			ObjectMeta: meta.ObjectMeta{
-				Name: "gateway",
-			},
-		},
-	}, gw.Listener{}, K8sListenerConfig{
-		Logger: hclog.NewNullLogger(),
-	})
-	canBind = listener.Bind(context.Background(), factory.NewRoute(&gw.HTTPRoute{
-		Spec: gw.HTTPRouteSpec{
-			CommonRouteSpec: gw.CommonRouteSpec{
-				ParentRefs: []gw.ParentRef{{
-					Name: "gateway",
-				}},
-			},
-		},
-	}))
-	require.True(t, canBind)
-
-	// not ready
-	listener = NewK8sListener(&K8sGateway{
-		Gateway: &gw.Gateway{
-			ObjectMeta: meta.ObjectMeta{
-				Name: "gateway",
-			},
-		},
-	}, gw.Listener{}, K8sListenerConfig{
-		Logger: hclog.NewNullLogger(),
-	})
-	listener.ListenerState.Status.Ready.Invalid = errors.New("invalid")
-	canBind = listener.Bind(context.Background(), factory.NewRoute(&gw.HTTPRoute{
-		Spec: gw.HTTPRouteSpec{
-			CommonRouteSpec: gw.CommonRouteSpec{
-				ParentRefs: []gw.ParentRef{{
-					Name: "gateway",
-				}},
-			},
-		},
-	}))
-	require.False(t, canBind)
-}
-
-func TestListenerCanBind_RouteKind(t *testing.T) {
-	t.Parallel()
-
-	routeMeta := meta.TypeMeta{}
-	routeMeta.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   gw.GroupVersion.Group,
-		Version: gw.GroupVersion.Version,
-		Kind:    "UDPRoute",
-	})
-	name := gw.SectionName("listener")
-
-	factory := NewFactory(FactoryConfig{
-		Logger: hclog.NewNullLogger(),
-	})
-
-	listener := NewK8sListener(&K8sGateway{
-		Gateway: &gw.Gateway{
-			ObjectMeta: meta.ObjectMeta{
-				Name: "gateway",
-			},
-		},
-	}, gw.Listener{
-		Protocol: gw.HTTPProtocolType,
-	}, K8sListenerConfig{
-		Logger: hclog.NewNullLogger(),
-	})
-	require.NoError(t, listener.Validate(context.Background(), nil, listener.gateway.Gateway, listener.listener))
-	canBind := listener.Bind(context.Background(), factory.NewRoute(&gw.UDPRoute{
-		TypeMeta: routeMeta,
-		Spec: gw.UDPRouteSpec{
-			CommonRouteSpec: gw.CommonRouteSpec{
-				ParentRefs: []gw.ParentRef{{
-					Name: "gateway",
-				}},
-			},
-		},
-	}))
-	require.False(t, canBind)
-
-	listener = NewK8sListener(&K8sGateway{
-		Gateway: &gw.Gateway{
-			ObjectMeta: meta.ObjectMeta{
-				Name: "gateway",
-			},
-		},
-	}, gw.Listener{
-		Name:     name,
-		Protocol: gw.HTTPProtocolType,
-	}, K8sListenerConfig{
-		Logger: hclog.NewNullLogger(),
-	})
-	canBind = listener.Bind(context.Background(), factory.NewRoute(&gw.UDPRoute{
-		TypeMeta: routeMeta,
-		Spec: gw.UDPRouteSpec{
-			CommonRouteSpec: gw.CommonRouteSpec{
-				ParentRefs: []gw.ParentRef{{
-					Name:        "gateway",
-					SectionName: &name,
-				}},
-			},
-		},
-	}))
-	require.False(t, canBind)
-}
-
-func TestListenerCanBind_AllowedNamespaces(t *testing.T) {
-	t.Parallel()
-
-	name := gw.SectionName("listener")
-	same := gw.NamespacesFromSame
-	selector := gw.NamespacesFromSelector
-	other := gw.Namespace("other")
-	routeMeta := meta.TypeMeta{}
-	routeMeta.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   gw.GroupVersion.Group,
-		Version: gw.GroupVersion.Version,
-		Kind:    "HTTPRoute",
-	})
-
-	factory := NewFactory(FactoryConfig{
-		Logger: hclog.NewNullLogger(),
-	})
-
-	listener := NewK8sListener(&K8sGateway{
-		Gateway: &gw.Gateway{
-			ObjectMeta: meta.ObjectMeta{
-				Name:      "gateway",
-				Namespace: "other",
-			},
-		},
-	}, gw.Listener{
-		Name:     name,
-		Protocol: gw.HTTPProtocolType,
-		AllowedRoutes: &gw.AllowedRoutes{
-			Namespaces: &gw.RouteNamespaces{
-				From: &same,
-			},
-		},
-	}, K8sListenerConfig{
-		Logger: hclog.NewNullLogger(),
-	})
-	canBind := listener.Bind(context.Background(), factory.NewRoute(&gw.HTTPRoute{
-		TypeMeta: routeMeta,
-		Spec: gw.HTTPRouteSpec{
-			CommonRouteSpec: gw.CommonRouteSpec{
-				ParentRefs: []gw.ParentRef{{
-					Name:        "gateway",
-					Namespace:   &other,
-					SectionName: &name,
-				}},
-			},
-		},
-	}))
-	require.False(t, canBind)
-	canBind = listener.Bind(context.Background(), factory.NewRoute(&gw.HTTPRoute{
-		TypeMeta: routeMeta,
-		Spec: gw.HTTPRouteSpec{
-			CommonRouteSpec: gw.CommonRouteSpec{
-				ParentRefs: []gw.ParentRef{{
-					Name:      "gateway",
-					Namespace: &other,
-				}},
-			},
-		},
-	}))
-	require.False(t, canBind)
-
-	listener = NewK8sListener(&K8sGateway{
-		Gateway: &gw.Gateway{
-			ObjectMeta: meta.ObjectMeta{
-				Name:      "gateway",
-				Namespace: "other",
-			},
-		},
-	}, gw.Listener{
-		Name:     name,
-		Protocol: gw.HTTPProtocolType,
-		AllowedRoutes: &gw.AllowedRoutes{
-			Namespaces: &gw.RouteNamespaces{
-				From: &selector,
-				Selector: &meta.LabelSelector{
-					MatchExpressions: []meta.LabelSelectorRequirement{{
-						Key:      "test",
-						Operator: meta.LabelSelectorOperator("invalid"),
-					}},
-				},
-			},
-		},
-	}, K8sListenerConfig{
-		Logger: hclog.NewNullLogger(),
-	})
-	canBind = listener.Bind(context.Background(), factory.NewRoute(&gw.HTTPRoute{
-		TypeMeta: routeMeta,
-		Spec: gw.HTTPRouteSpec{
-			CommonRouteSpec: gw.CommonRouteSpec{
-				ParentRefs: []gw.ParentRef{{
-					Name:        "gateway",
-					Namespace:   &other,
-					SectionName: &name,
-				}},
-			},
-		},
-	}))
-	require.False(t, canBind)
-	canBind = listener.Bind(context.Background(), factory.NewRoute(&gw.HTTPRoute{
-		TypeMeta: routeMeta,
-		Spec: gw.HTTPRouteSpec{
-			CommonRouteSpec: gw.CommonRouteSpec{
-				ParentRefs: []gw.ParentRef{{
-					Name:      "gateway",
-					Namespace: &other,
-				}},
-			},
-		},
-	}))
-	require.False(t, canBind)
-}
-
-func TestListenerCanBind_HostnameMatch(t *testing.T) {
-	t.Parallel()
-
-	factory := NewFactory(FactoryConfig{
-		Logger: hclog.NewNullLogger(),
-	})
-
-	routeMeta := meta.TypeMeta{}
-	routeMeta.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   gw.GroupVersion.Group,
-		Version: gw.GroupVersion.Version,
-		Kind:    "HTTPRoute",
-	})
-	name := gw.SectionName("listener")
-	hostname := gw.Hostname("hostname")
-	listener := NewK8sListener(&K8sGateway{
-		Gateway: &gw.Gateway{
-			ObjectMeta: meta.ObjectMeta{
-				Name: "gateway",
-			},
-		},
-	}, gw.Listener{
-		Name:     name,
-		Hostname: &hostname,
-		Protocol: gw.HTTPProtocolType,
-	}, K8sListenerConfig{
-		Logger: hclog.NewNullLogger(),
-	})
-	canBind := listener.Bind(context.Background(), factory.NewRoute(&gw.HTTPRoute{
-		TypeMeta: routeMeta,
-		Spec: gw.HTTPRouteSpec{
-			CommonRouteSpec: gw.CommonRouteSpec{
-				ParentRefs: []gw.ParentRef{{
-					Name:        "gateway",
-					SectionName: &name,
-				}},
-			},
-			Hostnames: []gw.Hostname{"other"},
-		},
-	}))
-	require.False(t, canBind)
-
-	canBind = listener.Bind(context.Background(), factory.NewRoute(&gw.HTTPRoute{
-		TypeMeta: routeMeta,
-		Spec: gw.HTTPRouteSpec{
-			CommonRouteSpec: gw.CommonRouteSpec{
-				ParentRefs: []gw.ParentRef{{
-					Name: "gateway",
-				}},
-			},
-			Hostnames: []gw.Hostname{"other"},
-		},
-	}))
-	require.False(t, canBind)
-}
-
-func TestListenerCanBind_NameMatch(t *testing.T) {
-	t.Parallel()
-
-	factory := NewFactory(FactoryConfig{
-		Logger: hclog.NewNullLogger(),
-	})
-
-	name := gw.SectionName("listener")
-	otherName := gw.SectionName("other")
-	listener := NewK8sListener(&K8sGateway{
-		Gateway: &gw.Gateway{
-			ObjectMeta: meta.ObjectMeta{
-				Name: "gateway",
-			},
-		},
-	}, gw.Listener{
-		Name:     name,
-		Protocol: gw.HTTPProtocolType,
-	}, K8sListenerConfig{
-		Logger: hclog.NewNullLogger(),
-	})
-	canBind := listener.Bind(context.Background(), factory.NewRoute(&gw.HTTPRoute{
-		Spec: gw.HTTPRouteSpec{
-			CommonRouteSpec: gw.CommonRouteSpec{
-				ParentRefs: []gw.ParentRef{{
-					Name:        "gateway",
-					SectionName: &otherName,
-				}},
-			},
-			Hostnames: []gw.Hostname{"other"},
-		},
-	}))
-	require.False(t, canBind)
 }
