+++ internal/k8s/reconciler/manager_test.go
@@ -91,7 +91,7 @@ func TestUpsertGateway(t *testing.T) {
 	require.Equal(t, expected, manager.UpsertGateway(context.Background(), inner))
 
 	client.EXPECT().PodWithLabels(gomock.Any(), gomock.Any()).Return(nil, nil).Times(2)
-	store.EXPECT().UpsertGateway(gomock.Any(), gomock.Any())
+	store.EXPECT().UpsertGateway(gomock.Any(), gomock.Any(), gomock.Any())
 	require.NoError(t, manager.UpsertGateway(context.Background(), inner))
 }
 
@@ -124,7 +124,7 @@ func TestUpsertHTTPRoute(t *testing.T) {
 	require.NoError(t, manager.UpsertHTTPRoute(context.Background(), &gw.HTTPRoute{}))
 
 	manager.namespaceMap[types.NamespacedName{Name: "gateway"}] = ""
-	store.EXPECT().UpsertRoute(gomock.Any(), gomock.Any()).Return(nil)
+	store.EXPECT().UpsertRoute(gomock.Any(), gomock.Any(), gomock.Any()).Return(nil)
 	require.NoError(t, manager.UpsertHTTPRoute(context.Background(), &gw.HTTPRoute{
 		Spec: gw.HTTPRouteSpec{
 			CommonRouteSpec: gw.CommonRouteSpec{
@@ -136,7 +136,7 @@ func TestUpsertHTTPRoute(t *testing.T) {
 	}))
 
 	client.EXPECT().IsManagedRoute(gomock.Any(), gomock.Any(), gomock.Any()).Return(true, nil)
-	store.EXPECT().UpsertRoute(gomock.Any(), gomock.Any()).Return(nil)
+	store.EXPECT().UpsertRoute(gomock.Any(), gomock.Any(), gomock.Any()).Return(nil)
 	require.NoError(t, manager.UpsertHTTPRoute(context.Background(), &gw.HTTPRoute{}))
 
 	client.EXPECT().IsManagedRoute(gomock.Any(), gomock.Any(), gomock.Any()).Return(true, nil)
@@ -174,7 +174,7 @@ func TestUpsertTCPRoute(t *testing.T) {
 	})
 
 	client.EXPECT().IsManagedRoute(gomock.Any(), gomock.Any(), gomock.Any()).Return(true, nil)
-	store.EXPECT().UpsertRoute(gomock.Any(), gomock.Any()).Return(nil)
+	store.EXPECT().UpsertRoute(gomock.Any(), gomock.Any(), gomock.Any()).Return(nil)
 	require.NoError(t, manager.UpsertTCPRoute(context.Background(), &gw.TCPRoute{}))
 }
 
@@ -194,7 +194,7 @@ func TestUpsertTLSRoute(t *testing.T) {
 	})
 
 	client.EXPECT().IsManagedRoute(gomock.Any(), gomock.Any(), gomock.Any()).Return(true, nil)
-	store.EXPECT().UpsertRoute(gomock.Any(), gomock.Any()).Return(nil)
+	store.EXPECT().UpsertRoute(gomock.Any(), gomock.Any(), gomock.Any()).Return(nil)
 	require.NoError(t, manager.UpsertTLSRoute(context.Background(), &gw.TLSRoute{}))
 }
