+++ internal/store/mocks/interfaces.go
@@ -36,6 +36,20 @@ func (m *MockStatusTrackingGateway) EXPECT() *MockStatusTrackingGatewayMockRecor
 	return m.recorder
 }
 
+// Bind mocks base method.
+func (m *MockStatusTrackingGateway) Bind(ctx context.Context, route store.Route) []string {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "Bind", ctx, route)
+	ret0, _ := ret[0].([]string)
+	return ret0
+}
+
+// Bind indicates an expected call of Bind.
+func (mr *MockStatusTrackingGatewayMockRecorder) Bind(ctx, route interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Bind", reflect.TypeOf((*MockStatusTrackingGateway)(nil).Bind), ctx, route)
+}
+
 // ID mocks base method.
 func (m *MockStatusTrackingGateway) ID() core.GatewayID {
 	m.ctrl.T.Helper()
@@ -78,34 +92,6 @@ func (mr *MockStatusTrackingGatewayMockRecorder) Meta() *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Meta", reflect.TypeOf((*MockStatusTrackingGateway)(nil).Meta))
 }
 
-// ShouldBind mocks base method.
-func (m *MockStatusTrackingGateway) ShouldBind(route store.Route) bool {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ShouldBind", route)
-	ret0, _ := ret[0].(bool)
-	return ret0
-}
-
-// ShouldBind indicates an expected call of ShouldBind.
-func (mr *MockStatusTrackingGatewayMockRecorder) ShouldBind(route interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ShouldBind", reflect.TypeOf((*MockStatusTrackingGateway)(nil).ShouldBind), route)
-}
-
-// ShouldUpdate mocks base method.
-func (m *MockStatusTrackingGateway) ShouldUpdate(other store.Gateway) bool {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ShouldUpdate", other)
-	ret0, _ := ret[0].(bool)
-	return ret0
-}
-
-// ShouldUpdate indicates an expected call of ShouldUpdate.
-func (mr *MockStatusTrackingGatewayMockRecorder) ShouldUpdate(other interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ShouldUpdate", reflect.TypeOf((*MockStatusTrackingGateway)(nil).ShouldUpdate), other)
-}
-
 // TrackSync mocks base method.
 func (m *MockStatusTrackingGateway) TrackSync(ctx context.Context, sync func() (bool, error)) error {
 	m.ctrl.T.Helper()
@@ -141,6 +127,20 @@ func NewMockGateway(ctrl *gomock.Controller) *MockGateway {
 // EXPECT returns an object that allows the caller to indicate expected use.
 func (m *MockGateway) EXPECT() *MockGatewayMockRecorder {
 	return m.recorder
+}
+
+// Bind mocks base method.
+func (m *MockGateway) Bind(ctx context.Context, route store.Route) []string {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "Bind", ctx, route)
+	ret0, _ := ret[0].([]string)
+	return ret0
+}
+
+// Bind indicates an expected call of Bind.
+func (mr *MockGatewayMockRecorder) Bind(ctx, route interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Bind", reflect.TypeOf((*MockGateway)(nil).Bind), ctx, route)
 }
 
 // ID mocks base method.
@@ -185,138 +185,6 @@ func (mr *MockGatewayMockRecorder) Meta() *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Meta", reflect.TypeOf((*MockGateway)(nil).Meta))
 }
 
-// ShouldBind mocks base method.
-func (m *MockGateway) ShouldBind(route store.Route) bool {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ShouldBind", route)
-	ret0, _ := ret[0].(bool)
-	return ret0
-}
-
-// ShouldBind indicates an expected call of ShouldBind.
-func (mr *MockGatewayMockRecorder) ShouldBind(route interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ShouldBind", reflect.TypeOf((*MockGateway)(nil).ShouldBind), route)
-}
-
-// ShouldUpdate mocks base method.
-func (m *MockGateway) ShouldUpdate(other store.Gateway) bool {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ShouldUpdate", other)
-	ret0, _ := ret[0].(bool)
-	return ret0
-}
-
-// ShouldUpdate indicates an expected call of ShouldUpdate.
-func (mr *MockGatewayMockRecorder) ShouldUpdate(other interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ShouldUpdate", reflect.TypeOf((*MockGateway)(nil).ShouldUpdate), other)
-}
-
-// MockRouteTrackingListener is a mock of RouteTrackingListener interface.
-type MockRouteTrackingListener struct {
-	ctrl     *gomock.Controller
-	recorder *MockRouteTrackingListenerMockRecorder
-}
-
-// MockRouteTrackingListenerMockRecorder is the mock recorder for MockRouteTrackingListener.
-type MockRouteTrackingListenerMockRecorder struct {
-	mock *MockRouteTrackingListener
-}
-
-// NewMockRouteTrackingListener creates a new mock instance.
-func NewMockRouteTrackingListener(ctrl *gomock.Controller) *MockRouteTrackingListener {
-	mock := &MockRouteTrackingListener{ctrl: ctrl}
-	mock.recorder = &MockRouteTrackingListenerMockRecorder{mock}
-	return mock
-}
-
-// EXPECT returns an object that allows the caller to indicate expected use.
-func (m *MockRouteTrackingListener) EXPECT() *MockRouteTrackingListenerMockRecorder {
-	return m.recorder
-}
-
-// CanBind mocks base method.
-func (m *MockRouteTrackingListener) CanBind(ctx context.Context, route store.Route) (bool, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "CanBind", ctx, route)
-	ret0, _ := ret[0].(bool)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// CanBind indicates an expected call of CanBind.
-func (mr *MockRouteTrackingListenerMockRecorder) CanBind(ctx, route interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CanBind", reflect.TypeOf((*MockRouteTrackingListener)(nil).CanBind), ctx, route)
-}
-
-// Config mocks base method.
-func (m *MockRouteTrackingListener) Config() store.ListenerConfig {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Config")
-	ret0, _ := ret[0].(store.ListenerConfig)
-	return ret0
-}
-
-// Config indicates an expected call of Config.
-func (mr *MockRouteTrackingListenerMockRecorder) Config() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Config", reflect.TypeOf((*MockRouteTrackingListener)(nil).Config))
-}
-
-// ID mocks base method.
-func (m *MockRouteTrackingListener) ID() string {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ID")
-	ret0, _ := ret[0].(string)
-	return ret0
-}
-
-// ID indicates an expected call of ID.
-func (mr *MockRouteTrackingListenerMockRecorder) ID() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ID", reflect.TypeOf((*MockRouteTrackingListener)(nil).ID))
-}
-
-// IsValid mocks base method.
-func (m *MockRouteTrackingListener) IsValid() bool {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "IsValid")
-	ret0, _ := ret[0].(bool)
-	return ret0
-}
-
-// IsValid indicates an expected call of IsValid.
-func (mr *MockRouteTrackingListenerMockRecorder) IsValid() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "IsValid", reflect.TypeOf((*MockRouteTrackingListener)(nil).IsValid))
-}
-
-// OnRouteAdded mocks base method.
-func (m *MockRouteTrackingListener) OnRouteAdded(route store.Route) {
-	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "OnRouteAdded", route)
-}
-
-// OnRouteAdded indicates an expected call of OnRouteAdded.
-func (mr *MockRouteTrackingListenerMockRecorder) OnRouteAdded(route interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "OnRouteAdded", reflect.TypeOf((*MockRouteTrackingListener)(nil).OnRouteAdded), route)
-}
-
-// OnRouteRemoved mocks base method.
-func (m *MockRouteTrackingListener) OnRouteRemoved(id string) {
-	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "OnRouteRemoved", id)
-}
-
-// OnRouteRemoved indicates an expected call of OnRouteRemoved.
-func (mr *MockRouteTrackingListenerMockRecorder) OnRouteRemoved(id interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "OnRouteRemoved", reflect.TypeOf((*MockRouteTrackingListener)(nil).OnRouteRemoved), id)
-}
-
 // MockListener is a mock of Listener interface.
 type MockListener struct {
 	ctrl     *gomock.Controller
@@ -338,21 +206,6 @@ func NewMockListener(ctrl *gomock.Controller) *MockListener {
 // EXPECT returns an object that allows the caller to indicate expected use.
 func (m *MockListener) EXPECT() *MockListenerMockRecorder {
 	return m.recorder
-}
-
-// CanBind mocks base method.
-func (m *MockListener) CanBind(ctx context.Context, route store.Route) (bool, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "CanBind", ctx, route)
-	ret0, _ := ret[0].(bool)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// CanBind indicates an expected call of CanBind.
-func (mr *MockListenerMockRecorder) CanBind(ctx, route interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CanBind", reflect.TypeOf((*MockListener)(nil).CanBind), ctx, route)
 }
 
 // Config mocks base method.
@@ -397,6 +250,18 @@ func (mr *MockListenerMockRecorder) IsValid() *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "IsValid", reflect.TypeOf((*MockListener)(nil).IsValid))
 }
 
+// RouteRemoved mocks base method.
+func (m *MockListener) RouteRemoved() {
+	m.ctrl.T.Helper()
+	m.ctrl.Call(m, "RouteRemoved")
+}
+
+// RouteRemoved indicates an expected call of RouteRemoved.
+func (mr *MockListenerMockRecorder) RouteRemoved() *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RouteRemoved", reflect.TypeOf((*MockListener)(nil).RouteRemoved))
+}
+
 // MockStatusTrackingRoute is a mock of StatusTrackingRoute interface.
 type MockStatusTrackingRoute struct {
 	ctrl     *gomock.Controller
@@ -420,20 +285,6 @@ func (m *MockStatusTrackingRoute) EXPECT() *MockStatusTrackingRouteMockRecorder 
 	return m.recorder
 }
 
-// Compare mocks base method.
-func (m *MockStatusTrackingRoute) Compare(other store.Route) store.CompareResult {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Compare", other)
-	ret0, _ := ret[0].(store.CompareResult)
-	return ret0
-}
-
-// Compare indicates an expected call of Compare.
-func (mr *MockStatusTrackingRouteMockRecorder) Compare(other interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Compare", reflect.TypeOf((*MockStatusTrackingRoute)(nil).Compare), other)
-}
-
 // ID mocks base method.
 func (m *MockStatusTrackingRoute) ID() string {
 	m.ctrl.T.Helper()
@@ -446,30 +297,6 @@ func (m *MockStatusTrackingRoute) ID() string {
 func (mr *MockStatusTrackingRouteMockRecorder) ID() *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ID", reflect.TypeOf((*MockStatusTrackingRoute)(nil).ID))
-}
-
-// OnBindFailed mocks base method.
-func (m *MockStatusTrackingRoute) OnBindFailed(err error, gateway store.Gateway) {
-	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "OnBindFailed", err, gateway)
-}
-
-// OnBindFailed indicates an expected call of OnBindFailed.
-func (mr *MockStatusTrackingRouteMockRecorder) OnBindFailed(err, gateway interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "OnBindFailed", reflect.TypeOf((*MockStatusTrackingRoute)(nil).OnBindFailed), err, gateway)
-}
-
-// OnBound mocks base method.
-func (m *MockStatusTrackingRoute) OnBound(gateway store.Gateway) {
-	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "OnBound", gateway)
-}
-
-// OnBound indicates an expected call of OnBound.
-func (mr *MockStatusTrackingRouteMockRecorder) OnBound(gateway interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "OnBound", reflect.TypeOf((*MockStatusTrackingRoute)(nil).OnBound), gateway)
 }
 
 // OnGatewayRemoved mocks base method.
@@ -533,20 +360,6 @@ func NewMockRoute(ctrl *gomock.Controller) *MockRoute {
 // EXPECT returns an object that allows the caller to indicate expected use.
 func (m *MockRoute) EXPECT() *MockRouteMockRecorder {
 	return m.recorder
-}
-
-// Compare mocks base method.
-func (m *MockRoute) Compare(other store.Route) store.CompareResult {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Compare", other)
-	ret0, _ := ret[0].(store.CompareResult)
-	return ret0
-}
-
-// Compare indicates an expected call of Compare.
-func (mr *MockRouteMockRecorder) Compare(other interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Compare", reflect.TypeOf((*MockRoute)(nil).Compare), other)
 }
 
 // ID mocks base method.
@@ -673,29 +486,29 @@ func (mr *MockStoreMockRecorder) Sync(ctx interface{}) *gomock.Call {
 }
 
 // UpsertGateway mocks base method.
-func (m *MockStore) UpsertGateway(ctx context.Context, gateway store.Gateway) error {
+func (m *MockStore) UpsertGateway(ctx context.Context, gateway store.Gateway, updateConditionFn func(store.Gateway) bool) error {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "UpsertGateway", ctx, gateway)
+	ret := m.ctrl.Call(m, "UpsertGateway", ctx, gateway, updateConditionFn)
 	ret0, _ := ret[0].(error)
 	return ret0
 }
 
 // UpsertGateway indicates an expected call of UpsertGateway.
-func (mr *MockStoreMockRecorder) UpsertGateway(ctx, gateway interface{}) *gomock.Call {
+func (mr *MockStoreMockRecorder) UpsertGateway(ctx, gateway, updateConditionFn interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "UpsertGateway", reflect.TypeOf((*MockStore)(nil).UpsertGateway), ctx, gateway)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "UpsertGateway", reflect.TypeOf((*MockStore)(nil).UpsertGateway), ctx, gateway, updateConditionFn)
 }
 
 // UpsertRoute mocks base method.
-func (m *MockStore) UpsertRoute(ctx context.Context, route store.Route) error {
+func (m *MockStore) UpsertRoute(ctx context.Context, route store.Route, updateConditionFn func(store.Route) bool) error {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "UpsertRoute", ctx, route)
+	ret := m.ctrl.Call(m, "UpsertRoute", ctx, route, updateConditionFn)
 	ret0, _ := ret[0].(error)
 	return ret0
 }
 
 // UpsertRoute indicates an expected call of UpsertRoute.
-func (mr *MockStoreMockRecorder) UpsertRoute(ctx, route interface{}) *gomock.Call {
+func (mr *MockStoreMockRecorder) UpsertRoute(ctx, route, updateConditionFn interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "UpsertRoute", reflect.TypeOf((*MockStore)(nil).UpsertRoute), ctx, route)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "UpsertRoute", reflect.TypeOf((*MockStore)(nil).UpsertRoute), ctx, route, updateConditionFn)
 }
